Finished part 2/3
The basis of this change relies around the fact that we're
passing a boolean to keep track of whether values should be
"scrubbed" or not.

To keep track of this, every time we recurse into another object
map, we check if the key for this object map is a scrubbable key.
If it is, it means all future values that are nested under this
key should be scrubbed.

When we come back up from our recurssion, we replace the scrubble
value we had found *before* we saw this key.

You can see this logic in the JSONObject constructor. The actual
scrubbing of keys happens in JSONTokener.nextValue().

This "should scrub" boolean is passed through the constructors
of the JSONArray and JSONObject objects, and is used whenever
we're reading a key when we do .nextValue().