Finished pat 3/3.
There are a few clever tricks on how this works.

1) We still are doing recursive calls here. So it's possible that
if an object is realllly deep, that we could run out of stack space.
There's noting from writing it in an iterative fashion. However,
to minimize code changes, I simply keep the recursive structure of
the calls, and simply didn't create objects.

2) The more exciting part is how we read in the string. Basically,
every time our "reader" returns a next() character, we write it
to our string builder. This has 2 cavaets.

  1) If we've already read the character before(we know this because
     the isUsingPrevious flag is set), then we don't write to to the
     stringBuffer.
  2) If we're in "scrubbing" mode, we don't write things to the
     the string buffer.

The "scrubbing" mode is used when we want to read characters in
from our reader, but not store them to the StringBuffer, until
after it's been scrubbed.

This code is about as secure as you can get, as we only store at
max 1 character of scrubale in memory at a time(besides what's
in the reader buffer).